requirements.txt
asgiref==3.2.10
Django==3.1.2
django-crispy-forms==1.12.0
django-filter==2.4.0
pytz==2021.1
sqlparse==0.4.1
